2023-11-speedrun/002/requirements.txt
fastapi==0.104.1
Jinja2==3.1.2
PyYAML==6.0.1
uvicorn[standard]==0.24.0